backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
langgraph
langchain
langchain-openai
openai
pydantic
python-jose[cryptography]
passlib[bcrypt]